Remove Encoder ID enum and instead just use the classes
Everything is unnecessarily complicated when the encoder ID enum is used
especially the implementation of new encoders. Without the enum it is now
possible to implement a new encoder without touching the source code at all
i.e. the whole tester can be bundled into a package for PIP but the user can
still implement a new encoder.