Remove sequences from Test and subtest Objects
It is not necessary to pass the sequences for Test and SubTest, since those
only initialize their respective data and encoding runs can be safely
initialized in the run_tests and compute_metrics functions since the
EncodingRun does not have any kind state. Additionally there is less
possibilities for the user to screw things up since they cannot edit the
sequences between passing them to Test and create_context.

One negative is that seeking and limiting frames might be more difficult.