Change how metrics are computed
With only CSV output there was no problem on determining if vmaf etc. needed
to be calculated but now since there are tables and csvs if user only needs
table output but the not outputted csv would include e.g. vmaf but the table
not. It would be computed unnecessarily wasting time.

Now the compute_metrics has an additional parameter for determining which
enable_fields should be looked at when computing metrics.

Overall the result is bit awkward but I think it is the best solutions that
is not massively over engineered

Update README.md