Improve file naming and Pool usage
Include all of resolution, bitdepth, chroma format, and fps in filename
because if the file is converted there would be collision with the original
file otherwise

Use `Pool.imap_unordered`. This should help with interrupting the testing
with KeyBoardInterrupt and should spread the encoding evenly instead of pre-
allocating them.